authenticate and authorise using api

+++ controllers/api/v1/posts_api.js
@@ -1,5 +1,6 @@
 const Post = require('../../../models/post');
 const Comment = require('../../../models/comment');
+const { post } = require('../../../routes/api/v1/posts');
 
 
 module.exports.index = async function(req , res){
@@ -19,13 +20,23 @@ module.exports.index = async function(req , res){
 
 module.exports.destroy = async function(req , res){
 try{
-let  posts = await Post.findById(req.params.id);
-   console.log('hello');
-    posts.remove();
-   await  Comment.deleteMany({post : req.params.id});   
-   return res.json(200,{
-     message: "Post and associated message deleted!"
-   });
+
+
+    if(post.user == req.user.id){
+      let  posts = await Post.findById(req.params.id);
+      console.log('hello');
+       posts.remove();
+      await  Comment.deleteMany({post : req.params.id});   
+      return res.json(200,{
+        message: "Post and associated message deleted!"
+      });
+    }
+    else{
+              return res.json(401 , {
+                             message:"you cannot delete this post."
+              });
+    }
+
 }
 catch(err){
      return res.json(500, {
